SEMANA15/exemplo-jwt/pythonProject/requirements.txt
fastapi
uvicorn
passlib[bcrypt]
python-jose[cryptography]
pydantic
motor  # Para trabalhar com MongoDB de forma assíncrona
python-multipart